We are Cure House, a company that’s passionately working towards building a healthier India.
We are Cure House, a company that’s passionately working towards
building a healthier India. It’s our staunch belief that, in this day
and age, a busy schedule shouldn’t get in the way of people getting
medical care on time. Our platform provides extensive support to you,
including the following: ‒ Find doctors in your locality ‒ Book
appointments with doctors online ‒ Order medicines online and get them
delivered to your door ‒ Receive alerts when it’s time to have
medicines, or when a consultation with a doctor is due We can't make
your schedule less demanding, but we sure can help you manage your
health better.